1560 Accept quill as param rather then bundle it as dependency
