+++ package.json
@@ -45,11 +45,12 @@
 		"eslint-plugin-import": "^2.2.0",
 		"eslint-plugin-jsx-a11y": "^4.0.0",
 		"eslint-plugin-react": "^6.10.3",
-		"webpack": "^2.3.3"
+		"webpack": "^2.3.3",
+		"raw-loader": "^0.5.1"
 	},
 	"dependencies": {
 		"lodash": "^4.17.4",
-		"quill": "^1.3.6",
-		"raw-loader": "^0.5.1"
+		"quill": "^1.3.6"
+
 	}
 }
